+++ ml_project/Makefile
@@ -33,6 +33,8 @@ data: requirements
 clean:
 	find . -type f -name "*.py[co]" -delete
 	find . -type d -name "__pycache__" -delete
+	find . -type d -name ".pytest_cache" -exec rm -rv {} +
+	find . -type d -name "outputs" -exec rm -rv {} +
 
 ## Lint using flake8
 lint:
