+++ ml_project/conf/transformer/transformer.yaml
@@ -1,0 +1,2 @@
+# @package _group_
+path: models/transformer.pkl
