+++ ml_project/ml_pipeline/src/entities/transformer_params.py
@@ -1,0 +1,6 @@
+from dataclasses import dataclass, MISSING
+
+
+@dataclass()
+class TransformerParams:
+    path: MISSING
