+++ ml_project/ml_pipeline/tests/test_predict_pipeline.py
@@ -2,8 +2,8 @@ import os
 import numpy as np
 from typing import NoReturn
 
-from src.predict_pipeline import predict_pipeline
-from src.entities.pipeline_params import PipelineParams
+from ..src.predict_pipeline import predict_pipeline
+from ..src.entities.pipeline_params import PipelineParams
 
 
 def test_predict_pipeline(pipeline_params: PipelineParams) -> NoReturn:
