+++ ml_project/ml_pipeline/tests/test_train_model.py
@@ -4,8 +4,8 @@ from sklearn.linear_model import LogisticRegression
 from sklearn.ensemble import RandomForestClassifier
 from sklearn.utils.validation import check_is_fitted
 
-from src.models.train_model import build_model, train_model
-from src.entities.model_params import ModelParams
+from ..src.models.train_model import build_model, train_model
+from ..src.entities.model_params import ModelParams
 
 
 def test_logistic_regression_build(params: ModelParams) -> NoReturn:
