+++ ml_project/requirements.txt
@@ -6,4 +6,4 @@ click==7.1.2
 pytest==6.2.3
 hydra_core==1.0.6
 omegaconf==2.0.6
-dataclasses
+dataclasses==0.6
